Use superscript for power instead of ^ notation
