+++ src/shared/lib/NumberUtils.ts
@@ -6,16 +6,3 @@ export function toConditionalPrecision(number: number, precision: number, thresh
 
     return number.toPrecision(precision);
 };
-
-export function toConditionalPrecisionWithMinimum(
-    positiveNumber:number,
-    precision:number,
-    precisionThreshold:number,
-    minimumExponent:number
-):string {
-    if (positiveNumber === 0 || Math.log10(positiveNumber) < minimumExponent) {
-        return `< 10^${minimumExponent}`;
-    } else {
-        return toConditionalPrecision(positiveNumber, precision, precisionThreshold);
-    }
-}
